add fine-grained engine configuration options
This replaces the previous system of passing configuration options
as parameters to the import task. Instead, omerta_logger now
includes a dedicated configuration file with far more detailed
configuration options allowing users to configure what data to
track on a per-model basis.

The import task now imports all enabled domains sequentially and
has a "daemon" mode which keeps running imports in an infinite loop.
This is a far better approach in terms of performance compared to
the previously recommended option of running cronjobs for regular
imports.